fix macos compiling compatible

+++ src/lib/explorer/extensions/commands/importkeyfile.cpp
@@ -68,7 +68,7 @@ console_result importkeyfile::invoke (std::ostream& output,
         root["unissued-asset-count"] += all_info.get_account_asset().size();
     } else {
         root["address-count"] = acc.get_hd_index();
-        root["unissued-asset-count"] = all_info.get_account_asset().size();
+        root["unissued-asset-count"] = static_cast<uint64_t>(all_info.get_account_asset().size());
     }
     output << root.toStyledString();
